dmeo launch updated and moved to control
