--- src/cougars_localization/launch/demo_launch.py
@@ -1,47 +1,0 @@
-import sys
-
-import launch
-import launch_ros.actions
-import launch_ros.descriptions
-
-import yaml
-
-def generate_launch_description():
-    '''
-    :author: Nelson Durrant
-    :date: September 2024
-    
-    Launches the sensor nodes for the vehicle.
-
-    :return: The launch description.
-    '''
-
-    namespace = ''
-    
-    for arg in sys.argv:
-        if arg.startswith('namespace:='):
-            namespace = arg.split(':=')[1]
-        if arg.startswith('param_file:='):
-            param_file = arg.split(':=')[1]
-
-    with open(param_file, 'r') as f:
-        vehicle_params = yaml.safe_load(f)
-    
-    return launch.LaunchDescription([
-        
-        # Serial Teensy connection
-        launch_ros.actions.Node(
-            package='fin_sub_cpp', 
-            executable='control_node', 
-            namespace=namespace,
-            parameters=[param_file],
-            output='log',
-        ),
-        # Start the control nodes
-        launch_ros.actions.Node(
-            package='cougars_control',
-            executable='coug_kinematics',
-            parameters=[param_file],
-            namespace=namespace,
-        ),
-    ])
